+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Version 0.5.0
+
+- Added ability to define column names, id column names, and table names
+TODO write tests!
+
 ## Version 0.4.0 (11-27-2014)
 
 - Added query builder functionality. Returned from `selectFrom`